refactor(frontend): Move MemberService to top AssignmentModule
