+++ frontend/src/app/assignment/modules/assignment/assignment.module.ts
@@ -26,7 +26,6 @@ import {SubmitModalComponent} from './submit-modal/submit-modal.component';
 import {AssignmentTasksComponent} from './tasks/tasks.component';
 import {StatisticsService} from "./statistics.service";
 import {SubmitService} from "./submit.service";
-import {MemberService} from "./member.service";
 import {UserModule} from "../../../user/user.module";
 
 @NgModule({
@@ -59,7 +58,6 @@ import {UserModule} from "../../../user/user.module";
     UserModule,
   ],
   providers: [
-    MemberService,
     StatisticsService,
     SubmitService,
   ],
